requirements.txt
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.26.4
jupyter>=1.0.0
ipython>=8.0.0
matplotlib>=3.7.0
litellm>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.0

